@@ -1,12 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.2"
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 8
-        targetSdkVersion 21
+        minSdkVersion 14
+        targetSdkVersion 28
         versionCode 12
         versionName "0.3.0"
     }
@@ -17,9 +16,9 @@ android {
 
 dependencies {
     //noinspection GradleDynamicVersion
-    compile 'com.android.support:support-annotations:+'
+    api 'androidx.annotation:annotation:1.1.0'
     //noinspection GradleDynamicVersion
-    provided 'com.android.support:support-v4:+'
+    compileOnly 'androidx.legacy:legacy-support-v4:1.0.0'
 }
 
 //apply from: 'https://raw.github.com/chrisbanes/gradle-mvn-push/master/gradle-mvn-push.gradle'
