@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.2"
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "rapid.decoder.sample"
-        minSdkVersion 8
-        targetSdkVersion 21
+        minSdkVersion 14
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -16,9 +15,9 @@ android {
 }
 
 dependencies {
-    compile project(':library')
-    compile project(':jpeg-decoder')
-    compile project(':png-decoder')
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:20.0.0'
+    implementation project(':library')
+    implementation project(':jpeg-decoder')
+    implementation project(':png-decoder')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.0.0'
 }
