@@ -1,6 +1,6 @@
 package rapid.decoder;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.io.InputStream;
 
