@@ -2,7 +2,7 @@ package rapid.decoder.binder;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public interface DrawableInflater {
     @Nullable
