@@ -1,6 +1,6 @@
 package rapid.decoder.cache;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.IOException;
 import java.io.OutputStream;
