@@ -3,8 +3,8 @@ package rapid.decoder.sample;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.Display;
 import android.view.LayoutInflater;
 import android.view.View;
