@@ -2,11 +2,13 @@ package rapid.decoder.sample;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.v4.app.ActionBarDrawerToggle;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.ActionBarActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.legacy.app.ActionBarDrawerToggle;
+import androidx.fragment.app.Fragment;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBar;
+
 import android.view.Gravity;
 import android.view.KeyEvent;
 import android.view.MenuItem;
@@ -18,7 +20,7 @@ import android.widget.ListView;
 
 import rapid.decoder.BitmapDecoder;
 
-public class MainActivity extends ActionBarActivity {
+public class MainActivity extends AppCompatActivity {
     public static final boolean TEST_BUILT_IN_DECODER = true;
 
     private ActionBarDrawerToggle mDrawerToggle;
